created seprate profile service for profile related fuctions
