+++ src/app/services/api.service.ts
@@ -3,6 +3,7 @@ import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http'
 import { Observable } from 'rxjs';
 import * as urlConfig from '../constants/url-config.json';
 import { environment } from 'src/assets/envirnoments/environment';
+import { ProfileService } from './profile.service';
 
 @Injectable({
   providedIn: 'root'
@@ -10,7 +11,6 @@ import { environment } from 'src/assets/envirnoments/environment';
 export class ApiService {
   public baseUrl:string=environment.surveyBaseURL;
   public token:string;
-  public profileData:any;
   public solutionId :any; 
   public entityType:any
   public userAuthToken:any=localStorage.getItem('accToken');
@@ -20,7 +20,7 @@ export class ApiService {
   public index:any;
   public fileSizeLimit:any;
 
-  constructor(private http:HttpClient) {}
+  constructor(private http:HttpClient, private profileService:ProfileService) {}
 
   get<T>(url: string, params?: HttpParams): Observable<T> {
     return this.http.get<T>(this.baseUrl+url, { params });
@@ -62,36 +62,8 @@ export class ApiService {
     }
     }
 
-  public get fetchProfileData(){
-    let rawProfileData = this.getRawProfileFromStorage();
-    return this.profileData = this.normalizeProfileData(rawProfileData);
-  }
-
-
-  public getRawProfileFromStorage(): any {
-    const data = localStorage.getItem('profileData');
-    if (!data) return null;
-    try {
-      return JSON.parse(data);
-    } catch (e) {
-      console.error('Failed to parse profileData from localStorage:', e);
-      return null;
-    }
-  }
-
-  public normalizeProfileData(profileData: any): any {
-    if (!profileData) return null;
-
-    const normalized: any = {};
-
-    Object.entries(profileData).forEach(([key, value]: [string, any]) => {
-      if (value && typeof value === 'object' && 'id' in value) {
-        normalized[key] = value.id;
-      } else {
-        normalized[key] = value;
-      }
-    });
-
-    return normalized;
+  public get profileData(){
+    let rawProfileData = this.profileService.getRawProfileFromStorage();
+    return this.profileService.normalizeProfileData(rawProfileData);
   }
 }
